Add files via upload
This is a python code which takes a document's json response from the firebase database and checks if in the admin device there is wallet of of the same id of the json response if such wallet is present then this code creates and a zip file containing a chraacter images and a text file with the character name and it's description or else , if the wallet folder is not present then a wallet folder is first created in which zip file of the character is created storing the details.  A wallet can have multiple zip files of characters.